rename & fix: 라운지 -> 메인, 메인 랜딩으로 디렉토리 변경 (#971)
* rename: main -> landing 디렉터리 변경

* rename: lounge -> main 디렉터리 변경

* fix: 행사 목록 클릭하면 행사로 넘어가도록 수정

* fix: in progress check 20자 넘칠 때 깨지지 않게 설정
