+++ client/cypress/e2e/createEventByGuest.cy.ts
@@ -9,7 +9,7 @@ beforeEach(() => {
 
 describe('Flow: 비회원이 랜딩 페이지에서부터 이벤트를 생성 완료하는 flow', () => {
   it('비회원이 랜딩페이지에서 "정산 시작하기" 버튼을 누르면 로그인 페이지로 이동해야 한다.', () => {
-    cy.visit(ROUTER_URLS.main);
+    cy.visit(ROUTER_URLS.landing);
     cy.get('button').contains('정산 시작하기').click();
     cy.url().should('include', ROUTER_URLS.login);
   });
