+++ client/src/components/Design/components/CreatedEventItem/CreatedEventView.tsx
@@ -1,9 +1,10 @@
 /** @jsxImportSource @emotion/react */
+import {useNavigate} from 'react-router-dom';
+
 import {useTheme} from '@components/Design/theme/HDesignProvider';
 import {CreatedEvent} from 'types/serviceType';
 
-import Flex from '../Flex/Flex';
-import Text from '../Text/Text';
+import {TextButton, Flex, Text} from '@components/Design';
 
 import {inProgressCheckStyle} from './CreatedEventItem.style';
 
@@ -12,20 +13,28 @@ function InProgressCheck({inProgress}: {inProgress: boolean}) {
 
   return (
     <div css={inProgressCheckStyle({theme, inProgress})}>
-      <Text size="tiny" className="in-progress-check-text">
+      <Text size="tiny" className="in-progress-check-text" css={{whiteSpace: 'nowrap'}}>
         {inProgress ? '정산 중' : '정산 완료'}
       </Text>
     </div>
   );
 }
 
-export const CreatedEventView = ({eventName, isFinished}: CreatedEvent) => {
+export const CreatedEventView = ({eventName, isFinished, eventId}: CreatedEvent) => {
+  const navigate = useNavigate();
+
   return (
-    <Flex gap="0.5rem" minHeight="2.5rem" alignItems="center">
+    <Flex
+      width="100%"
+      gap="0.5rem"
+      minHeight="2.5rem"
+      alignItems="center"
+      onClick={() => navigate(`/event/${eventId}/admin`)}
+    >
       <InProgressCheck inProgress={!isFinished} />
-      <Text size="bodyBold" color="onTertiary">
+      <TextButton textSize="bodyBold" textColor="onTertiary">
         {eventName}
-      </Text>
+      </TextButton>
     </Flex>
   );
 };
