+++ client/src/pages/landing/Nav/Nav.tsx
@@ -3,11 +3,11 @@ import {useNavigate} from 'react-router-dom';
 import {User} from 'types/serviceType';
 import {IconHeundeut} from '@components/Design/components/Icons/Icons/IconHeundeut';
 
-import {Button, Text, TopNav, IconButton} from '@HDesign/index';
+import {Button, TopNav} from '@HDesign/index';
 
 import {ROUTER_URLS} from '@constants/routerUrls';
 
-import {navFixedStyle, navStyle, navWrapperStyle} from './Nav.style';
+import {navFixedStyle, navWrapperStyle} from './Nav.style';
 
 type NavProps = Pick<User, 'isGuest'>;
