+++ client/src/pages/landing/Section/CreatorSection/Avatar.style.ts
