+++ client/src/pages/landing/Section/DescriptionSection/DescriptionSection.style.ts
