+++ client/src/pages/landing/Section/DescriptionSection/DescriptionSection.tsx
