+++ client/src/pages/landing/Section/DescriptionSection/index.ts
