+++ client/src/pages/landing/Section/FeatureSection/AutoCalculate/AutoCalculate.style.ts
