+++ client/src/pages/landing/Section/FeatureSection/AutoCalculate/AutoCalculate.tsx
