+++ client/src/pages/landing/Section/FeatureSection/SimpleShare/SimpleShare.style.ts
