+++ client/src/pages/landing/Section/FeatureSection/SimpleShare/SimpleShare.tsx
