+++ client/src/pages/landing/Section/FeatureSection/SimpleShare/index.ts
