+++ client/src/pages/landing/Section/MainSection/MainSection.style.ts
