+++ client/src/pages/landing/Section/MainSection/MainSection.tsx
