+++ client/src/pages/landing/Section/MainSection/index.ts
