--- client/src/pages/main/index.ts
@@ -1,1 +1,0 @@
-export {default as MainPage} from './MainPage';
