+++ client/src/router.tsx
@@ -4,9 +4,12 @@ import {lazy, Suspense} from 'react';
 import {ROUTER_URLS} from '@constants/routerUrls';
 
 import App from './App';
+const LandingPage = lazy(() => import('@pages/landing/LandingPage'));
+const MainPage = lazy(() => import('@pages/main/MainPage'));
+
 const UserInfoLoader = lazy(() => import('@components/Loader/UserInfo/UserInfoLoader'));
-const EditUserAccountPage = lazy(() => import('@pages/lounge/edit-account/EditUserAccountPage'));
-const EditUserNicknamePage = lazy(() => import('@pages/lounge/edit-nickname/EditUserNicknamePage'));
+const EditUserAccountPage = lazy(() => import('@pages/main/edit-account/EditUserAccountPage'));
+const EditUserNicknamePage = lazy(() => import('@pages/main/edit-nickname/EditUserNicknamePage'));
 const ErrorPage = lazy(() => import('@pages/fallback/ErrorPage'));
 const SendErrorPage = lazy(() => import('@pages/fallback/SendErrorPage'));
 const CreateGuestEventFunnel = lazy(() => import('@pages/event/create/guest/CreateGuestEventFunnel'));
@@ -18,7 +21,6 @@ const EventLoader = lazy(() => import('@components/Loader/EventData/EventDataLoa
 const AuthGate = lazy(() => import('@pages/event/[eventId]/admin/AuthGate'));
 const EventPageLayout = lazy(() => import('@pages/event/[eventId]/EventPageLayout'));
 const SendPage = lazy(() => import('@pages/event/[eventId]/home/send/SendPage'));
-const MainPage = lazy(() => import('@pages/main/MainPage'));
 const HomePage = lazy(() => import('@pages/event/[eventId]/home/HomePage'));
 const AdminPage = lazy(() => import('@pages/event/[eventId]/admin/AdminPage'));
 const AddBillFunnel = lazy(() => import('@pages/event/[eventId]/admin/add-bill/AddBillFunnel'));
@@ -31,10 +33,9 @@ const QRCodePage = lazy(() => import('@pages/event/[eventId]/qrcode/QRCodePage')
 const LoginPage = lazy(() => import('@pages/login/LoginPage'));
 const SettingPage = lazy(() => import('@pages/setting/SettingPage'));
 const WithdrawPage = lazy(() => import('@pages/setting/withdraw/WithdrawPage'));
-const LoungePage = lazy(() => import('@pages/lounge/Lounge'));
 const LoginRedirectPage = lazy(() => import('@pages/login/LoginRedirectPage'));
 const LoginFailFallback = lazy(() => import('@pages/login/LoginFailFallback'));
-const CreatedEventsPage = lazy(() => import('@pages/lounge/events/CreatedEventsPage'));
+const CreatedEventsPage = lazy(() => import('@pages/main/events/CreatedEventsPage'));
 const EventPageLoading = lazy(() => import('@pages/fallback/EventPageLoading'));
 const EditEventName = lazy(() => import('@pages/event/[eventId]/admin/edit-event-name/EditEventNamePage'));
 const BillDetailPage = lazy(() => import('@pages/event/[eventId]/home/bill-detail/BillDetailPage'));
@@ -50,8 +51,8 @@ const router = createBrowserRouter([
     children: [
       {
         index: true,
-        path: ROUTER_URLS.main,
-        element: <MainPage />,
+        path: ROUTER_URLS.landing,
+        element: <LandingPage />,
       },
       {
         path: ROUTER_URLS.createGuestEvent,
@@ -115,7 +116,7 @@ const router = createBrowserRouter([
         ],
       },
       {
-        path: ROUTER_URLS.lounge,
+        path: ROUTER_URLS.main,
         element: (
           <Suspense>
             <UserInfoLoader />
@@ -124,7 +125,7 @@ const router = createBrowserRouter([
         children: [
           {
             index: true,
-            element: <LoungePage />,
+            element: <MainPage />,
           },
           {
             path: ROUTER_URLS.editUserAccount,
